requirements.txt
requests
requests-oauthlib
pandas
python-dotenv
matplotlib
seaborn